projects/chestxray/requirements.txt
cytoolz
dask
google-cloud==0.32.0
numpy
pandas
Pillow
scipy
scikit-image
tensorflow>=1.7
tensorflow-gpu
tensorflow-hub
tflearn
tqdm
toolz
tabulate
tqdm

